requirements.txt
PyQt5==5.15.10
loguru==0.7.2
python-dotenv==1.0.1
requests==2.32.3
pyaudio==0.2.14
markdown2==2.4.13
anthropic==0.40.0


